[CUBRIDQA-53] introduce requirement checking

+++ CTP/ha_repl/src/com/navercorp/cubridqa/ha_repl/dispatch/Dispatch.java
@@ -185,7 +185,8 @@ public class Dispatch {
 
 		File[] subList = new File(context.getCurrentLogDir()).listFiles();
 		for (File file : subList) {
-			file.delete();
+			if (file.isFile())
+				file.delete();
 		}
 
 		allFile = new File(getFileNameForDispatchAll());
